Almost have a contribution framework working
