+++ .gitignore
@@ -1,6 +1,7 @@
 .venv
 .ipynb_checkpoints
-*/__pycache__
+__pycache__/
+**/__pycache__/
 */*.pyc
 */*.pyo
 */*.pyd
