+++ session2.md
@@ -1,0 +1,11 @@
+In this session, we undertook a comprehensive review and enhancement of a Python-based backtesting framework for Bitcoin trading strategies. The project, designed for extensibility and performance, initially supported both serial (bar-by-bar) and parallel (vectorized) evaluation of trading strategies, with a focus on long-only, all-in/all-out logic. The user's primary goals were to speed up backtests, ensure robust and correct metrics, and align the results of vectorized and serial strategy implementations.
+
+We began by examining the Buy & Hold and SMA Crossover strategies, identifying that the Buy & Hold implementation was not correctly tracking the BTC price in the equity curve. This was traced to the way signals were converted to positions in the backtest engine: the use of a simple assignment meant the position reverted to cash after the first bar, rather than remaining long. We fixed this by introducing a cumulative sum for Buy & Hold, and later, a state machine for all strategies to ensure correct all-in/all-out transitions.
+
+Attention then shifted to the SMA Crossover strategy, where the vectorized (parallel) and serial (bar-by-bar) implementations produced different results. The root cause was a subtle lookahead bias in the vectorized approach: signals were being acted upon at the same bar they were generated, rather than at the next bar as in the serial version. This led to discrepancies in performance metrics such as Sharpe ratio, total return, and win rate. We resolved this by shifting the vectorized signals forward by one bar, ensuring that trades were only executed using information available at the close of the previous bar.
+
+Throughout the session, we also addressed several technical issues, including dtype warnings in pandas, proper handling of open positions at the end of the backtest, and robust trade record generation. The metrics module was improved for numerical stability and correct index handling, ensuring that drawdown and rolling Sharpe calculations were accurate and aligned with the equity curve.
+
+The session included iterative testing and visualization in Jupyter notebooks, with the user providing feedback on the accuracy of equity curves, drawdowns, and buy/sell points. Each round of changes was validated by re-running the backtests and comparing the results of the Buy & Hold and SMA Crossover strategies, both visually and through summary statistics.
+
+By the end of the session, the backtesting framework was significantly improved: it now supports efficient, vectorized evaluation of strategies without sacrificing correctness or introducing lookahead bias. The Buy & Hold and SMA Crossover strategies both produce realistic, robust results, and the framework is well-positioned for further strategy development and research. The careful alignment of serial and parallel logic ensures that future strategies can be implemented and tested with confidence in the accuracy of their performance metrics.
