--- src/backtest.py
@@ -1,129 +1,0 @@
-import pandas as pd
-import numpy as np
-from metrics import (
-    sharpe_ratio, total_return, n_trades, win_rate, max_drawdown, annualized_return,
-    rolling_sharpe_ratio, unrealized_drawdown_series, realized_drawdown_series
-)
-
-def run_backtest(strategy_class, df, initial_capital=10000):
-    # Local copy in memory for us to work with
-    df = df.copy()    
-
-    # Initialize strategy
-    strat = strategy_class(initial_capital=initial_capital)
-
-    # Get all signals at once
-    signals = strat.get_signals(df)
-    
-    # State machine for position: 0 = out, 1 = long
-    positions = pd.Series(0.0, index=df.index)
-    current_position = 0.0
-    for i, signal in enumerate(signals):
-        if current_position == 0.0:
-            if signal == 'buy':
-                current_position = 1.0
-        elif current_position == 1.0:
-            if signal == 'sell':
-                current_position = 0.0
-        positions.iloc[i] = current_position
-    
-    # Calculate position changes
-    position_changes = positions.diff().fillna(positions.iloc[0])
-    
-    # Identify entry and exit points
-    entries = position_changes == 1
-    exits = position_changes == -1
-    
-    # Initialize equity curve and coin holdings with proper dtypes
-    equity_curve = pd.Series(initial_capital, index=df.index, dtype=float)
-    coins = pd.Series(0.0, index=df.index, dtype=float)
-    
-    # Track current position and coins
-    current_position = 0
-    current_coins = 0
-    
-    # Process each bar
-    for i in range(len(df)):
-        if entries.iloc[i]:
-            # Enter position
-            current_position = 1
-            current_coins = equity_curve.iloc[i-1] / df['close'].iloc[i]
-            coins.iloc[i:] = current_coins
-        elif exits.iloc[i]:
-            # Exit position
-            current_position = 0
-            current_coins = 0
-            coins.iloc[i:] = 0
-        
-        # Update equity curve
-        if current_position == 1:
-            equity_curve.iloc[i] = current_coins * df['close'].iloc[i]
-        else:
-            equity_curve.iloc[i] = equity_curve.iloc[i-1] if i > 0 else initial_capital
-    
-    # Generate trade records
-    trades = []
-    entry_idx = None
-    entry_price = None
-    
-    for i in range(len(df)):
-        if entries.iloc[i]:
-            entry_idx = i
-            entry_price = df['close'].iloc[i]
-        elif exits.iloc[i] and entry_idx is not None:
-            exit_price = df['close'].iloc[i]
-            pnl = (exit_price - entry_price) * coins.iloc[entry_idx]
-            
-            trades.append({
-                'entry': df['time'].iloc[entry_idx],
-                'entry_idx': entry_idx,
-                'entry_price': entry_price,
-                'exit': df['time'].iloc[i],
-                'exit_idx': i,
-                'exit_price': exit_price,
-                'pnl': pnl
-            })
-            entry_idx = None
-            entry_price = None
-    
-    # Always close any open position at the end
-    if entry_idx is not None or current_position == 1:
-        if entry_idx is None:
-            # Find the last entry point if we're in position but don't have an entry record
-            entry_idx = entries[entries].index[-1] if any(entries) else 0
-            entry_price = df['close'].iloc[entry_idx]
-        
-        exit_price = df['close'].iloc[-1]
-        pnl = (exit_price - entry_price) * coins.iloc[entry_idx]
-        
-        trades.append({
-            'entry': df['time'].iloc[entry_idx],
-            'entry_idx': entry_idx,
-            'entry_price': entry_price,
-            'exit': df['time'].iloc[-1],
-            'exit_idx': len(df) - 1,
-            'exit_price': exit_price,
-            'pnl': pnl
-        })
-        
-        # Update final equity
-        equity_curve.iloc[-1] = initial_capital + sum(t['pnl'] for t in trades)
-    
-    # Calculate returns (hourly)
-    returns = equity_curve.pct_change().fillna(0).values
-
-    # Metrics
-    results = {
-        'sharpe': sharpe_ratio(returns),
-        'total_return': total_return(equity_curve),
-        'n_trades': n_trades(trades),
-        'win_rate': win_rate(trades),
-        'max_drawdown': max_drawdown(equity_curve),
-        'annualized_return': annualized_return(equity_curve),
-        'equity_curve': equity_curve,
-        'trades': trades,
-        'rolling_sharpe': rolling_sharpe_ratio(equity_curve),
-        'unrealized_drawdown': unrealized_drawdown_series(equity_curve),
-        'realized_drawdown': realized_drawdown_series(equity_curve, trades),
-    }
-    return results
